[19.0][MIG] website_event_questions_by_ticket: Migration to 19.0
