+++ website_event_questions_by_ticket/README.rst
@@ -21,13 +21,13 @@ Conditional Events Questions
     :target: http://www.gnu.org/licenses/lgpl-3.0-standalone.html
     :alt: License: LGPL-3
 .. |badge3| image:: https://img.shields.io/badge/github-OCA%2Fevent-lightgray.png?logo=github
-    :target: https://github.com/OCA/event/tree/18.0/website_event_questions_by_ticket
+    :target: https://github.com/OCA/event/tree/19.0/website_event_questions_by_ticket
     :alt: OCA/event
 .. |badge4| image:: https://img.shields.io/badge/weblate-Translate%20me-F47D42.png
-    :target: https://translation.odoo-community.org/projects/event-18-0/event-18-0-website_event_questions_by_ticket
+    :target: https://translation.odoo-community.org/projects/event-19-0/event-19-0-website_event_questions_by_ticket
     :alt: Translate me on Weblate
 .. |badge5| image:: https://img.shields.io/badge/runboat-Try%20me-875A7B.png
-    :target: https://runboat.odoo-community.org/builds?repo=OCA/event&target_branch=18.0
+    :target: https://runboat.odoo-community.org/builds?repo=OCA/event&target_branch=19.0
     :alt: Try me on Runboat
 
 |badge1| |badge2| |badge3| |badge4| |badge5|
@@ -74,7 +74,7 @@ Bug Tracker
 Bugs are tracked on `GitHub Issues <https://github.com/OCA/event/issues>`_.
 In case of trouble, please check there if your issue has already been reported.
 If you spotted it first, help us to smash it by providing a detailed and welcomed
-`feedback <https://github.com/OCA/event/issues/new?body=module:%20website_event_questions_by_ticket%0Aversion:%2018.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**>`_.
+`feedback <https://github.com/OCA/event/issues/new?body=module:%20website_event_questions_by_ticket%0Aversion:%2019.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**>`_.
 
 Do not contact contributors directly about support or help with technical issues.
 
@@ -96,6 +96,10 @@ Contributors
   - Carolina Fernandez
   - Pilar Vargas
 
+- `Acsone <https://www.acsone.eu>`__:
+
+  - Maxime Franco
+
 Maintainers
 -----------
 
@@ -109,6 +113,6 @@ OCA, or the Odoo Community Association, is a nonprofit organization whose
 mission is to support the collaborative development of Odoo features and
 promote its widespread use.
 
-This module is part of the `OCA/event <https://github.com/OCA/event/tree/18.0/website_event_questions_by_ticket>`_ project on GitHub.
+This module is part of the `OCA/event <https://github.com/OCA/event/tree/19.0/website_event_questions_by_ticket>`_ project on GitHub.
 
 You are welcome to contribute. To learn how please visit https://odoo-community.org/page/Contribute.
