+++ website_event_questions_by_ticket/models/__init__.py
@@ -1,3 +1,4 @@
 # License LGPL-3.0 or later (http://www.gnu.org/licenses/lgpl).
 
 from . import event_event
+from . import event_question
