+++ website_event_questions_by_ticket/models/event_event.py
@@ -1,7 +1,7 @@
 # Copyright 2019 Tecnativa - Pedro M. Baeza
 # License LGPL-3.0 or later (http://www.gnu.org/licenses/lgpl).
 
-from odoo import fields, models
+from odoo import models
 
 
 class EventEvent(models.Model):
@@ -23,14 +23,3 @@ class EventEvent(models.Model):
                 or (set(ticket_ids) & set(x.restricted_ticket_ids.ids))
             )
         )
-
-
-class EventQuestion(models.Model):
-    _inherit = "event.question"
-
-    restricted_ticket_ids = fields.Many2many(
-        comodel_name="event.event.ticket",
-        string="Limited to tickets",
-        domain="[('event_id', '=', parent.id)]",
-        ondelete="restrict",
-    )
