+++ website_event_questions_by_ticket/views/event_event_views.xml
@@ -2,24 +2,6 @@
 <!-- Copyright 2019 Tecnativa - Pedro M. Baeza
      License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl). -->
 <odoo>
-    <record model="ir.ui.view" id="view_event_question_form">
-        <field name="model">event.event</field>
-        <field name="inherit_id" ref="website_event.event_event_view_form" />
-        <field name="arch" type="xml">
-            <xpath
-                expr="//field[@name='question_ids']/form/sheet/notebook"
-                position="before"
-            >
-                <group>
-                    <field
-                        name="restricted_ticket_ids"
-                        widget="many2many_tags"
-                        placeholder="Leave it empty for being available for all tickets"
-                    />
-                </group>
-            </xpath>
-        </field>
-    </record>
     <record model="ir.ui.view" id="event_event_view_form">
         <field name="model">event.event</field>
         <field name="inherit_id" ref="website_event.event_event_view_form" />
@@ -30,6 +12,11 @@
             >
                 <field name="restricted_ticket_ids" widget="many2many_tags" />
             </xpath>
+            <field name="question_ids" position="attributes">
+                <attribute
+                    name="context"
+                >{'list_view_ref': 'event.event_question_view_list_add', 'form_view_ref': 'website_event_questions_by_ticket.event_question_view_form'}</attribute>
+            </field>
         </field>
     </record>
 </odoo>
