+++ website_event_questions_by_ticket/views/event_question_views.xml
@@ -1,0 +1,23 @@
+<?xml version="1.0" encoding="utf-8" ?>
+<!-- Copyright 2019 Tecnativa - Pedro M. Baeza
+     License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl). -->
+<odoo>
+    <record model="ir.ui.view" id="event_question_view_form">
+        <field name="name">event.question.view.list.add</field>
+        <field name="model">event.question</field>
+        <field name="inherit_id" ref="event.event_question_view_form" />
+        <field name="mode">primary</field>
+        <field name="priority" eval="20" />
+        <field name="arch" type="xml">
+            <xpath expr="//notebook" position="before">
+                <group>
+                    <field
+                        name="restricted_ticket_ids"
+                        widget="many2many_tags"
+                        placeholder="Leave it empty for being available for all tickets"
+                    />
+                </group>
+            </xpath>
+        </field>
+    </record>
+</odoo>
